Sort imports: config
See
https://github.com/SexualHealthInnovations/callisto-core/issues/4#issuecomment-223693086.

This contains the configuration I will use to sort imports with `isort`.
In the next commit I'll actually sort the imports.

I also added a note to `CONTRIBUTING.md` about adhering to the django
coding style guide. @kelseyq, you may want to check that the comment is where you
think it should be. I didn't think we needed to provide any more detail
than simply linking to the django guide, but if you'd like more I'm
happy to add it.