Fix PEP8 violations: config
Here I prepare to fix PEP8 violations with the help of the [python-mode vim
plugin](https://github.com/klen/python-mode), which uses
[autopep8](https://pypi.python.org/pypi/autopep8) by including necessary
configuration and dependencies.

I included these dependencies in requirements_dev.txt on the assumption that
contributors to `callisto-core` would be using them to ensure pep8 and
django coding style compliance as they work (but that users of the
project would have their own solution for style compliance).