Automatic fix of PEP8 issues
This commit represents a completely automatic fix of PEP8 issues with
[autopep8](https://pypi.python.org/pypi/autopep8) using this command in
the root directory of the project:

```
autopep8 --in-place --recursive --aggressive --aggressive .
```

@kelseyq [prefers automatic checking and
fixing](https://github.com/SexualHealthInnovations/callisto-core/pull/64#issuecomment-230057261);
therefore, I haven't manually tweaked anything that `autopep8` did. I'm
submitting this for review to see what y'all think, and I'm also going
to look through the diff later to see if there's anything `autopep8` did
that seems undesirable (though @kelseyq is of course the only
authority on what's desirable or not in this project).